requirements.txt
httpx==0.28.1
pydantic==2.11.3
PyYAML==6.0.2
pytest==8.3.5
ruff==0.11.6
tenacity==9.1.2
